Initial commit: re-create registration form

+++ service-logic/src/main/java/com/itechart/service/dto/PriceDto.java
@@ -15,7 +15,7 @@ public class PriceDto {
     private double furnitureAndCoatingsCleaning;
     private double industrialCleaning;
     private double poolCleaning;
-    private BigDecimal smallRoom;
-    private BigDecimal bigRoom;
-    private BigDecimal bathroom;
+    private double smallRoom;
+    private double bigRoom;
+    private double bathroom;
 }
